requirements.txt
fastapi>=0.70.0
uvicorn[standard]>=0.15.0
tensorflow>=2.9.0
pillow>=8.0.0
numpy>=1.19.0
streamlit>=1.0
python-multipart>=0.0.5
requests>=2.25.0
pytest>=6.0
